Almost done :)

+++ app/database/migrations/2013_12_26_185448_create_settings_table.php
@@ -17,7 +17,7 @@ class CreateSettingsTable extends Migration {
 
 			$table->integer('user_id');
 			$table->string('font_face');
-			$table->integer('font_size');
+			$table->float('font_size');
 			$table->integer('font_weight');
 		});
 	}
